Bug fix: extra letters being set to yellow.
If the letter appeared in the word but was not greened, ALL instances of the letter were being set to yellow.
Now, only the amount of letters in the answer are set to yellow.
E.g. for answer 'BALLS', if 'LLXXL' is guessed, only the first 2 'L's will be set to yellow. The third will be greyed.